backend/pyproject.toml
[project]
name = "symbolic-breeder-backend"
version = "0.1.0"
description = "Backend for Symbolic Breeder - evolutionary programming tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
include = ["app*", "seeds*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
